Journal: Change slot from u64 to usize
This avoids the need for many `@intCast()`s to accomodate 32-bit builds.
Also, the slot is logically an index type, so this makes more sense semantically.

+++ src/vsr/journal.zig
@@ -60,7 +60,7 @@ comptime {
 /// - `journal.faulty`
 ///
 /// A header's slot is `header.op % constants.journal_slot_count`.
-const Slot = struct { index: u64 };
+const Slot = struct { index: usize };
 
 /// An inclusive, non-empty range of slots.
 const SlotRange = struct {
